File APK ASSIGNMENT - Add Breed Name
Continue your LAB_WEEK_05 project. Remember that you defined the “breeds”
variable in ImageData.kt. Currently it’s only showing the URL of the loaded image.
Change the URL Text to the breed of the current Cat Image.
Note: Due to the incompleteness of the Cat API, not all cats have the Breed data, so
instead you can display “Unknown” for the ones that are not found.